Lecture 10 - Restoring heap functionality

+++ src/memory/heap/heap.c
@@ -3,6 +3,7 @@
 #include "status.h"
 #include "memory/memory.h"
 #include <stdbool.h>
+#include <stdint.h>
 
 static int heap_validate_table(void* ptr, void* end, struct heap_table* table)
 {
@@ -22,7 +23,7 @@ out:
 
 static bool heap_validate_alignment(void* ptr)
 {
-    return ((unsigned int)ptr % PEACHOS_HEAP_BLOCK_SIZE) == 0;
+    return ((uintptr_t)ptr % PEACHOS_HEAP_BLOCK_SIZE) == 0;
 }
 
 int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* table)
